Cell Autosizing: changed supported fonts, improved/added cell height calculation. Renamed namespace Width to Size. / fixed test namespace
